Move fleetlock components to internal package
* fleetlock is an application, not intended to be used
as a library
