+++ internal/decode.go
